SKIF-316 - TestdataService.getTestNumber_0() endres til TestdataServer.getTestNumber0() SKIF-317 - W2DAdapterProxyHandler og D2WAdapterProxyHandler mister generisk typeinfo for adaptormetode SKIF-318 - DefaultTypeMapper rapportere feil når returtype på metode innholder TypeVariable SKIF-319 - StoreServiceWSI manger metode getObjectsIgnoreMissing
[git-p4: depot-paths = "//skif/SkifKode/trunk/": change = 49473]

+++ skif-store-test/skif-store-test-api/src/main/java/no/statkart/skif/storetest/domain/basic/BubbleWithRelation.java
@@ -19,7 +19,7 @@ import no.statkart.skif.storetest.domain.mockup.FooId;
 public class BubbleWithRelation extends AbstractStoreTestBubble {
     private static final long serialVersionUID = 1L;
 
-    private long nr;
+    private int nr;
     private String text;
     private SimpleId<?> simpleId;
 
@@ -35,11 +35,11 @@ public class BubbleWithRelation extends AbstractStoreTestBubble {
         return (BubbleWithRelationId<?>) super.getId();
     }
 
-    public long getNr() {
+    public int getNr() {
         return nr;
     }
 
-    public void setNr(long nr) {
+    public void setNr(int nr) {
         this.nr = nr;
     }
