+++ skif-store-test/skif-store-test-api/src/main/java/no/statkart/skif/storetest/domain/basic/HistWithRelation.java
@@ -13,7 +13,7 @@ import no.statkart.skif.storetest.domain.StoreTestBubble;
 public class HistWithRelation extends AbstractStoreTestBubbleWithHistory {
     private static final long serialVersionUID = 1L;
 
-    private long nr;
+    private int nr;
     private String text;
     private int testSetNumber;
     private HistSimpleId<?> histSimpleId;
@@ -30,11 +30,11 @@ public class HistWithRelation extends AbstractStoreTestBubbleWithHistory {
         return (HistWithRelationId<?>) super.getId();
     }
 
-    public long getNr() {
+    public int getNr() {
         return nr;
     }
 
-    public void setNr(long nr) {
+    public void setNr(int nr) {
         this.nr = nr;
     }
